nb2/pyproject.toml
[tool.poetry]
name = "nb2"
version = "0.1.0"
description = "nb2"
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.7.3"
nonebot2 = "^2.0.0-beta.1"

[tool.poetry.dev-dependencies]
nb-cli = "^0.6.0"

[tool.nonebot]
plugins = ["nonebot_plugin_crazy_thursday", "nonebot_plugin_wordle", "nonebot_plugin_emojimix", "nonebot_plugin_drawer"]
plugin_dirs = ["src/plugins"]

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
